server/app/requirements.txt
distribute==0.6.34
MySQL-python==1.2.4
argparse==1.2.1
wsgiref==0.1.2
Django==1.5.1
PIL==1.1.7
PyYAML==3.10
South==0.7.6
Werkzeug==0.8.3
argh==0.23.1
chromelogger==0.3.0
django-allauth==0.10.1
django-analytical==0.15.0
django-extensions==1.1.1
django-filter==0.6
djangorestframework==2.3.2
facebook-sdk==0.4.0
gunicorn==0.17.4
httplib2==0.8
jsonpickle==0.4.0
meld3==0.6.10
nodeenv==0.6.5
oauth2==1.5.211
oauthlib==0.4.0
pathtools==0.1.2
python-openid==2.2.5
requests==1.2.0
requests-oauthlib==0.3.1
simplejson==3.2.0
six==1.3.0
supervisor==3.0b1
watchdog==0.6.0
